+++ src/assets/colors.ts
@@ -1,0 +1,7 @@
+export const COLORS = {
+    white: '#FFF',
+    gray: 'gray',
+
+    mainGradientTop: '#dddddd',
+    mainGradientBottom: '#7e7e7e',
+}