+++ src/assets/screenBreakpoints.ts
@@ -1,0 +1,10 @@
+import { IScreenBreakpoints } from "../interfaces/screenBreakpoints.interface";
+
+export const screenBreakpoints: IScreenBreakpoints = {
+    mobileXS: '576px',
+    mobileSM: '768px',
+    tabletsMD: '992px',
+    laptopsLG: '1200px',
+    desktopXL: '1400px',
+    tvXXL: '1920px'
+}