+++ src/components/Text.tsx
@@ -1,0 +1,11 @@
+import styled from 'styled-components';
+import { COLORS } from '../assets/colors';
+
+const Text = styled.text`
+    font-size: 20px;
+    color: ${COLORS.white};
+
+    font-family: Verdana, Geneva, Tahoma, sans-serif
+`
+
+export default Text;