+++ src/index.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client'
 
-import Home from './pages/home/index'
+import Home from './pages/Home'
+import './styles/global.css'
 
 const root = ReactDOM.createRoot(document.querySelector('#root') as Element);