+++ src/interfaces/view.interface.ts
@@ -1,0 +1,7 @@
+import { IChildrenProps } from "./childrenProps.interface"
+
+export interface IViewProps extends IChildrenProps {
+    height?: string;
+    width?: string;
+    flexDirection?: string;
+}